+++ src/plugins/piral-page-layouts/src/create.ts
@@ -1,8 +1,8 @@
 import * as actions from './actions';
-import { ComponentType, createElement, PropsWithChildren } from 'react';
-import { useLocation } from 'react-router-dom';
-import { PiralPlugin, PageComponentProps, Dict, GlobalState, withApi, defaultRender, useGlobalState, RouteSwitchProps } from 'piral-core';
-import type { PiletPageLayoutsApi, PageLayoutRegistration } from './types';
+import { ComponentType } from 'react';
+import { PiralPlugin, PageComponentProps, withApi } from 'piral-core';
+import { getPageLayouts, withPageLayouts } from './utils';
+import type { PiletPageLayoutsApi } from './types';
 
 /**
  * Available configuration options for the page layout plugin.
@@ -17,48 +17,6 @@ export interface PageLayoutsConfig {
    * Defines the initially available layouts that can be used.
    */
   layouts?: Record<string, ComponentType<PageComponentProps>>;
-}
-
-function getPageLayouts(items: Record<string, ComponentType<PageComponentProps>>) {
-  const layouts: Dict<PageLayoutRegistration> = {};
-
-  if (items && typeof items === 'object') {
-    Object.keys(items).forEach((name) => {
-      layouts[name] = {
-        pilet: undefined,
-        component: items[name],
-      };
-    });
-  }
-
-  return layouts;
-}
-
-const DefaultLayout: React.FC<PropsWithChildren> = props => defaultRender(props.children);
-
-function createPageWrapper(Routes: ComponentType<RouteSwitchProps>, fallback = 'default'): ComponentType<RouteSwitchProps> {
-  return (props) => {
-    const location = useLocation();
-    const data = props.paths.find(m => m.matcher.test(location.pathname));
-    const layout = data?.meta?.layout || fallback;
-    const registration = useGlobalState((s) => s.registry.pageLayouts[layout] || s.registry.pageLayouts[fallback]);
-    const Layout = registration?.component || DefaultLayout;
-    return createElement(Layout, props, createElement(Routes, props));
-  };
-}
-
-function withPageLayouts(pageLayouts: Dict<PageLayoutRegistration>, fallback: string) {
-  return (state: GlobalState): GlobalState => ({
-    ...state,
-    components: {
-      ...state.components,
-      RouteSwitch: createPageWrapper(state.components.RouteSwitch, fallback),
-    },
-    registry: {
-      ...state.registry,
-      pageLayouts,
-    },
-  });
 }
 
 /**
